The binary font uses pipes and underscores, four rows high and three columns wide. what else does it do? It "looks like" a readable number.
We're going to have to pass a variety of specs including the fact that the OCR recognizes numbers like zero, one, two, three, four, five, six, seven, eight, nine, ten, garble, 110101100, garble mixed in, 1234567890, and 123 456 789.

`git push origin` and `git push origin master` do the same thing.

Although we "can't read the template literal directly", what we can do is look at its decomposed form .split('').

Also like how this is supposed to be sequential. We don't see the results clogging up the console, of logging multiple test "specs" we just log one. The first one is to log zero.

This brings up the question of computing the equality of arrays. We want to compare our input array describing the zero, that is each character of the "new font zero" that is making this possible.

The formatting for this commit was made possible by VS Code, in which you can indent blocks of code to the left even when they're already at the left, and only for those lines that are at the left.