Looks like a use case for React. Wish the https://app.codesubmit.io/c/votingworks/8f6cdf3d-da8b-4b5a-9e93-909da8ad34c1/6c1fa646-c068-435f-bb40-812a1eee1e2e url updated the commits pushed to master without "having" to refresh the page! What's the difference between plagiarism and online resources? We already know. https://flexiple.com/javascript/javascript-array-equality/
Nothing was explained and accounted for.

It looks like JSON.stringify works!

If we don't want to skip, we can remove the "x" in front of the tests, with some find and replace. "What" if we want to clear up to certain server outputs like...   > 153 |     )).toBe('123,456,789');
        |        ^
    154 |   });
    155 | });
    156 |

    at Object.<anonymous> (ocr-numbers.spec.js:153:8)?

There's a comma after ' _ \n'
    + ' _|\n'
    + '|_ \n'
    + '   ', out of habit, per the CodeSignal Team, in ocr-numbers.spec.js!

With a switch statement and the specs we can pass the code by direct "pattern matching" among the input.